`RowColIterator` definition
